use longs for large hex constant literals, this avoids scary looking warnings on earlier versions of python
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@758389 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/python/qpid/spec010.py
@@ -484,7 +484,7 @@ class Spec(Node):
     buffer: direct("vbin32"),
     int: direct("int64"),
     long: direct("int64"),
-    float: direct("float"),
+    float: direct("double"),
     None.__class__: direct("void"),
     list: direct("list"),
     tuple: direct("list"),
